4.22-1

+++ INCLUDE/RES.h
@@ -4,6 +4,6 @@ int people_max(struct GameInfo *pg,char *path);
 void people_cost(int *data,struct GameInfo *pg);
 int happiness_count(struct GameInfo *pg);
 int happiness_rate_count(struct GameInfo *pg);
-int constrc_count(struct GameInfo *pg,char *path);
+long int constrc_count(struct GameInfo *pg,char *path);
 
 #endif
